requirements.txt
Flask==3.0.3
gunicorn==22.0.0
openai==1.58.1
pinecone-client[grpc]==5.0.1
Werkzeug==3.0.3
